+++ src/AppInstallerSharedLib/Public/winget/Runtime.h
@@ -51,6 +51,10 @@ namespace AppInstaller::Runtime
     //  3. the token is not already elevated
     bool IsRunningWithLimitedToken();
 
+    // Determines if the given amount of stack bytes are available.
+    // If the answer cannot be determined properly, the return value will be `false`.
+    DECLSPEC_NOINLINE bool IsStackAvailable(size_t bytes);
+
     // Returns true if this is a release build; false if not.
     inline constexpr bool IsReleaseBuild()
     {